+++ CHANGELOG.md
@@ -27,6 +27,7 @@ Support for Python 3.3 has been dropped, it reached its end of life.
   This allows shorter code in the bots, as an 'overwrite' configuration parameter can be directly passed to the function.
 - Bots can specify a static method `check(parameters)` which can perform individual checks specific to the bot.
   These functions will be called by `intelmqctl check` if the bot is configured with the given parameters
+- lib/bot: top level bot parameters (description, group, module, name) are exposed as members of the class.
 
 ### Bots
 #### Collectors